Extract to interface.
